# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestAll API",
        "description": "REST API documentation for the RestAll restaurant management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/menu/items": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get all menu items",
                "operationId": "c78c6b66bab20e84dc26153850b07bdb",
                "responses": {
                    "200": {
                        "description": "List of menu items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "menu_category_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Caesar Salad"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Classic salad with croutons",
                                                "nullable": true
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 12.5
                                            },
                                            "is_available": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Salads"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create a new menu item",
                "operationId": "9374dbcb03a1d95b44c044365c362763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "menu_category_id",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "menu_category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Greek Salad"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fresh vegetables with feta cheese"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menu/items/{id}": {
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update an existing menu item",
                "operationId": "7f7d319df20e1bd44928786c25540fde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "menu_category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Greek Salad Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fresh vegetables with extra feta"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 11.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete a menu item",
                "operationId": "1196abdaefe1321a0f5e7e5566f97fd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Menu item deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menu/items/{id}/availability": {
            "patch": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Toggle menu item availability (stop-list)",
                "operationId": "56a36461770f91e1bd30475cc8491640",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Availability toggled successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token obtained during login",
                "scheme": "bearer"
            }
        }
    },
...
}